most of the create event styles have now been converted to makestyles
