+++ src/App.css
@@ -398,289 +398,3 @@ a {
     font-size: 1.6rem;
     width: 90%;
 }
-
-/* FormPageThree three styles */
-
-.formPageThreeContainer {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-}
-
-@media (max-width: 1278px) {
-    .formPageThreeContainer {
-        width: 85%;
-        align-items: center;
-        justify-content: center;
-    }
-}
-
-.formPageThreeContainer h3 {
-    text-align: left;
-    width: 100%;
-    font-size: 1.8rem;
-    font-weight: 500;
-    color: #1a0f2c;
-    margin: 5px 0;
-}
-
-@media (max-width: 1278px) {
-    .formPageThreeContainer h3 {
-        margin-left: 4%;
-    }
-}
-
-@media (max-width: 600px) {
-    @media (max-width: 1278px) {
-        .formPageThreeContainer h3 {
-            margin-left: 0;
-        }
-    }
-}
-
-.formPageThreeCardContainter {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    border: 2px solid #f3f3f3;
-    box-shadow: 0px 4px 15px rgba(179, 179, 179, 0.1);
-    border-radius: 25px;
-    margin-top: 10px;
-    padding: 30px;
-    width: 90%;
-}
-
-@media (max-width: 800px) {
-    .formPageThreeCardContainter {
-        width: 90%;
-    }
-}
-
-.pageThreeCard {
-    display: flex;
-    width: 100%;
-    min-height: 200px;
-}
-
-@media (max-width: 800px) {
-    .pageThreeCard {
-        flex-direction: column;
-        align-items: center;
-    }
-}
-
-.pageThreeImg {
-    max-width: 50%;
-    border: 8px solid #58d473;
-    border-radius: 25px;
-    max-height: 300px;
-}
-
-@media (max-width: 800px) {
-    .pageThreeImg {
-        max-width: 90%;
-    }
-}
-
-.pageThreeText {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-    margin-left: 30px;
-    text-align: left;
-    width: 50%;
-}
-
-@media (max-width: 800px) {
-    .pageThreeText {
-        width: 82%;
-        margin-top: 10px;
-        height: 100px;
-    }
-}
-
-@media (max-width: 695px) {
-    .pageThreeText {
-        width: 96%;
-    }
-}
-
-/* .modifierContainer {
-    display: flex;
-    flex-wrap: wrap;
-    width: 100%;
-} */
-
-/* .individualModifierContainer {
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    width: 45%;
-    margin-top: 30px;
-    margin-left: 3%;
-}
-
-@media (max-width: 600px) {
-    .individualModifierContainer {
-        margin-top: 25px;
-        width: 100%;
-    }
-} */
-
-/* page four style */
-
-.formPageFourContainer {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-}
-
-@media (max-width: 1278px) {
-    .formPageFourContainer {
-        justify-content: center;
-        align-items: center;
-        width: 90%;
-        margin-left: 5%;
-    }
-}
-
-.pageFourCard {
-    display: flex;
-    width: 100%;
-    min-height: 200px;
-}
-
-@media (max-width: 600px) {
-    .pageFourCard {
-        flex-direction: column;
-        align-items: center;
-    }
-}
-
-.pageFourImg {
-    max-width: 50%;
-    border: 8px solid #58d473;
-    border-radius: 25px;
-    max-height: 300px;
-}
-
-@media (max-width: 600px) {
-    .pageFourImg {
-        max-width: 90%;
-    }
-}
-
-.pageFourText {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-    margin-left: 30px;
-    text-align: left;
-    width: 50%;
-}
-
-@media (max-width: 600px) {
-    .pageFourText {
-        width: 96%;
-        margin-top: 10px;
-        height: 100px;
-    }
-}
-
-.searchFriendsBtn {
-    background: #82df96;
-    padding: 15px 20px;
-    border-radius: 5px;
-    margin-left: 5%;
-    color: white;
-    outline: none;
-    font-size: 2rem;
-    font-weight: bold;
-    width: 15%;
-}
-
-@media (max-width: 800px) {
-    .searchFriendsBtn {
-        width: 25%;
-    }
-}
-
-.pageFourText {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-    margin-left: 30px;
-    text-align: left;
-    width: 50%;
-}
-
-@media (max-width: 600px) {
-    .pageFourText {
-        width: 96%;
-        margin-top: 10px;
-        height: 120px;
-    }
-}
-
-/* create event button div */
-/* .createFormButtonDiv {
-    display: flex;
-    justify-content: flex-end;
-    margin-right: 5%;
-    margin-bottom: 20px;
-    width: 100%;
-}
-
-@media (max-width: 800px) {
-    .createFormButtonDiv {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        width: 100%;
-        margin-right: 0;
-    }
-}
-
-.createRightBtn {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border: 2px solid #e5e5e5;
-    background: rgba(255, 255, 255, 0.5);
-    padding: 15px 20px;
-    border-radius: 15px;
-    margin: 20px 20px;
-    color: #b3b3b3;
-    outline: none;
-    font-size: 2rem;
-    font-weight: bold;
-    width: 15%;
-}
-
-@media (max-width: 800px) {
-    .createRightBtn {
-        width: 30%;
-    }
-}
-
-.createLeftBtn {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background: #58d473;
-    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.05);
-    padding: 15px 20px;
-    border-radius: 15px;
-    margin: 20px 0;
-    color: white;
-    outline: none;
-    font-size: 2rem;
-    font-weight: bold;
-    width: 15%;
-}
-
-@media (max-width: 800px) {
-    .createLeftBtn {
-        width: 30%;
-    }
-} */
