+++ src/components/create-events/form-container/form-page-four/FormPageFour.js
@@ -4,7 +4,15 @@ import SearchFriends from './search-friends/SearchFriends.js';
 
 const FormPageFour = () => {
     return (
-        <div className="formPageFourContainer">
+        <div
+            style={{
+                display: 'flex',
+                flexDirection: 'column',
+                justifyContent: 'center',
+                alignItems: 'center',
+                width: '100%',
+            }}
+        >
             <EventCard />
             <SearchFriends />
         </div>
